Move component tests alongside sources
