+++ frontend/src/components/areas/private/components/session/__tests__/login-form.test.jsx
@@ -6,7 +6,7 @@ import React from 'react';
 import { render, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import { BrowserRouter } from 'react-router-dom';
-import LoginForm from '../../src/components/areas/private/components/session/login-form';
+import LoginForm from '../login-form';
 
 // Instantiate router context
 const router = {
