+++ frontend/src/components/areas/public/features/task/components/__tests__/task-solution-drawer.test.js
@@ -4,7 +4,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
-import TaskSolutionDrawer from '../../../src/components/areas/public/features/task/components/send-solution-drawer';
+import TaskSolutionDrawer from '../send-solution-drawer';
 import { BrowserRouter } from 'react-router-dom';
 import { debug } from 'jest-preview';
